replaces URLConnection based http client with AdvancedHttpClient
