+++ pom.xml
@@ -44,12 +44,6 @@
 
   <dependencies>
 
-    <dependency>
-      <groupId>org.codehaus.jackson</groupId>
-      <artifactId>jackson-mapper-asl</artifactId>
-      <version>1.9.2</version>
-    </dependency>
-
     <!-- test scope -->
 
     <dependency>
