+++ src/main/java/sonia/scm/webhook/WebHookModule.java
@@ -30,8 +30,7 @@ import com.google.inject.AbstractModule;
 
 import org.mapstruct.factory.Mappers;
 import sonia.scm.plugin.Extension;
-import sonia.scm.webhook.impl.JacksonWebHookMarshaller;
-import sonia.scm.webhook.impl.URLWebHookHttpClient;
+import sonia.scm.webhook.impl.AhcWebHookHttpClient;
 import sonia.scm.webhook.internal.WebHookMapper;
 
 /**
@@ -50,8 +49,7 @@ public class WebHookModule extends AbstractModule
   protected void configure()
   {
     bind(WebHookContext.class);
-    bind(WebHookMarshaller.class).to(JacksonWebHookMarshaller.class);
-    bind(WebHookHttpClient.class).to(URLWebHookHttpClient.class);
+    bind(WebHookHttpClient.class).to(AhcWebHookHttpClient.class);
     bind(WebHookMapper.class).to(Mappers.getMapper(WebHookMapper.class).getClass());
   }
 }
