+++ src/main/java/sonia/scm/webhook/impl/AhcWebHookHttpClient.java
@@ -1,0 +1,83 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020-present Cloudogu GmbH and Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package sonia.scm.webhook.impl;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import sonia.scm.net.ahc.AdvancedHttpClient;
+import sonia.scm.net.ahc.AdvancedHttpRequestWithBody;
+import sonia.scm.net.ahc.AdvancedHttpResponse;
+import sonia.scm.webhook.HttpMethod;
+import sonia.scm.webhook.WebHookHttpClient;
+
+import javax.inject.Inject;
+import java.io.IOException;
+
+public class AhcWebHookHttpClient implements WebHookHttpClient {
+
+  private static final Logger LOG = LoggerFactory.getLogger(AhcWebHookHttpClient.class);
+
+  private final AdvancedHttpClient client;
+
+  @Inject
+  public AhcWebHookHttpClient(AdvancedHttpClient client) {
+    this.client = client;
+  }
+
+  @Override
+  public void execute(HttpMethod method, String url) throws IOException {
+    execute(method, url, null);
+  }
+
+  @Override
+  public void execute(HttpMethod method, String url, Object data) throws IOException {
+    HttpMethod m = method;
+
+    if (m == HttpMethod.AUTO) {
+      if (data != null) {
+        m = HttpMethod.POST;
+      } else {
+        m = HttpMethod.GET;
+      }
+    }
+
+    AdvancedHttpRequestWithBody request = client.method(m.name(), url);
+    if (data != null) {
+      request.jsonContent(data);
+    }
+
+    handleRespone(url, request.request());
+  }
+
+  private void handleRespone(String url, AdvancedHttpResponse response) {
+    int statusCode = response.getStatus();
+
+    if ((statusCode >= 200) && (statusCode < 300)) {
+      LOG.info("webhook {} ended successfully with status code {}", url, statusCode);
+    } else {
+      LOG.warn("webhook {} failed with statusCode {}", url, statusCode);
+    }
+  }
+
+}
